Various improvements
Added name of card a dice is rolled on to the output.
Added new parameters to onCardMove event.
Split the donate code so that the card controller is the only one to
manipulate the markers on the card.
Remove the spurious "and" as a trait choice in the askChoice dialog.